feat: Create line equation for P, K and MO
Now users are able to choose between a fixed value to multiply the input value of P, K and MO or a line equation.
Also changed the configuration file structure.
Fixed a bug with main window when try to close all pages in one click.
Replace all insert of QLineEdit to setText